+++ App.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, {useState} from 'react';
 import { StyleSheet, Text, View, Image, TouchableOpacity } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
@@ -8,11 +8,17 @@ import Home from './Home.js';
 import { createMaterialBottomTabNavigator } from '@react-navigation/material-bottom-tabs';
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
 import FontAwesome from 'react-native-vector-icons/FontAwesome';
+import Login from './Login.js';
 
 // import Icons from "./Icons";
 const Tab = createMaterialBottomTabNavigator();
 
 export default function App() {
+
+const [userLoggedIn, setUserLoggedIn] = useState(false);
+
+if(userLoggedIn){
+
   return (
     <NavigationContainer>
       <Tab.Navigator
@@ -54,10 +60,14 @@ export default function App() {
     </NavigationContainer>
   );
 
-
+      
 }
+ else {
+  return (<Login/>); 
+}
+} 
 
 
 const styles = StyleSheet.create({
   
-});
+});;
