+++ Login.js
@@ -1,0 +1,40 @@
+import React from "react";
+import { StyleSheet, Text, View, Image, Alert,   } from 'react-native';
+import { Card, ListItem, Button, Icon } from 'react-native-elements';
+import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons';
+import { SafeAreaView, TextInput } from "react-native";
+
+const Login = () => {
+  const [text, onChangeText] = React.useState("Enter Phone Number");
+  const [number, onChangeNumber] = React.useState(null);
+
+
+    return(
+      <SafeAreaView>
+      <TextInput
+        style={styles.input}
+        onChangeText={onChangeText}
+        value={text}
+      />
+      <TextInput
+        style={styles.input}
+        onChangeText={onChangeNumber}
+        value={number}
+        placeholder="Phone Number"
+        keyboardType="numeric"
+      />
+    </SafeAreaView>
+  );
+};
+
+
+const styles = StyleSheet.create({
+  input: {
+    height: 40,
+    margin: 12,
+    borderWidth: 1,
+    padding: 10,
+  },
+});
+  
+  export default Login;
